requirements.txt
Django==1.9.9
django-autoslug==1.9.3
django-wysiwyg-redactor==0.4.9.1
Pillow==3.4.2
python-decouple==3.0
